+++ seqkit/cmd/seq.go
@@ -74,7 +74,6 @@ var seqCmd = &cobra.Command{
 		rna2dna := getFlagBool(cmd, "rna2dna")
 		color := getFlagBool(cmd, "color")
 		validateSeq := getFlagBool(cmd, "validate-seq")
-		validateSeqLength := getFlagValidateSeqLength(cmd, "validate-seq-length")
 		minLen := getFlagInt(cmd, "min-len")
 		maxLen := getFlagInt(cmd, "max-len")
 		qBase := getFlagPositiveInt(cmd, "qual-ascii-base")
@@ -112,8 +111,6 @@ var seqCmd = &cobra.Command{
 		}
 
 		seq.ValidateSeq = validateSeq
-		seq.ValidateWholeSeq = false
-		seq.ValidSeqLengthThreshold = validateSeqLength
 		seq.ValidSeqThreads = config.Threads
 		seq.ComplementThreads = config.Threads
 
@@ -446,7 +443,6 @@ func init() {
 	seqCmd.Flags().BoolP("rna2dna", "", false, "RNA to DNA")
 	seqCmd.Flags().BoolP("color", "k", false, "colorize sequences - to be piped into \"less -R\"")
 	seqCmd.Flags().BoolP("validate-seq", "v", false, "validate bases according to the alphabet")
-	seqCmd.Flags().IntP("validate-seq-length", "V", 10000, "length of sequence to validate (0 for whole seq)")
 	seqCmd.Flags().IntP("min-len", "m", -1, "only print sequences longer than or equal to the minimum length (-1 for no limit)")
 	seqCmd.Flags().IntP("max-len", "M", -1, "only print sequences shorter than or equal to the maximum length (-1 for no limit)")
 	seqCmd.Flags().IntP("qual-ascii-base", "b", 33, "ASCII BASE, 33 for Phred+33")